# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wheel fortune doc API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/material-thing": {
            "get": {
                "tags": [
                    "Вещи"
                ],
                "summary": "Получение списка вещей",
                "operationId": "53182061aec9bb8bf72386e16e5dd057",
                "responses": {
                    "200": {
                        "description": "ОК",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MaterialThing"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": {
                            "role": "admin"
                        }
                    }
                ]
            },
            "post": {
                "tags": [
                    "Вещи"
                ],
                "summary": "Создание вещи",
                "operationId": "7460baaa36a6e7b992356ad1e9ddca1c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "count"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Блокнот А5"
                                    },
                                    "count": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "ОК",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MaterialThing"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": {
                            "role": "admin"
                        }
                    }
                ]
            }
        },
        "/api/material-thing/{id}": {
            "get": {
                "tags": [
                    "Вещи"
                ],
                "summary": "Получение вещи по id",
                "operationId": "5e4c7213feb4e5678bcd2bf945856c28",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id вещи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ОК",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MaterialThing"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ничего не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Вещи с таким id не существует"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": {
                            "role": "admin"
                        }
                    }
                ]
            },
            "put": {
                "tags": [
                    "Вещи"
                ],
                "summary": "Обновление вещи  по id",
                "operationId": "4da5f791e2f9ab47db7561cbafbff914",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id вещи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Блокнот А5"
                                    },
                                    "count": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ОК",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MaterialThing"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ничего не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Вещи с таким id не существует"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": {
                            "role": "admin"
                        }
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Вещи"
                ],
                "summary": "Удаление вещи по id",
                "operationId": "78d11539508be32b509403d3a3a54e65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id вещи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ОК",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Вещь успешна удалена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Действие запрещено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Этот приз удалить нельзя, его выйграли"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ничего не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Вещи с таким id не существует"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": {
                            "role": "admin"
                        }
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MaterialThing": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Блокнот А5"
                    },
                    "count": {
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}